Bug #1438
« Previous | 1 of 11 | Next »
Regress #339: 01_common_tasks
Regress #462: Time synchronization between system components
Incorrect timezone in file title.csv

<==== Mercurial data ====>
Branch: master

+++ common/src/utils/common/util.h
@@ -160,7 +160,7 @@ public:
 /*
 * Returns current time zone offset in seconds
 */
-qint16 currentTimeZone();
+int currentTimeZone();
 
 
 static const qint64 UTC_TIME_DETECTION_THRESHOLD = 1000000ll * 3600*24*100;
